Minimal copypasta into gradle plugin
